# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 库存调拨规划器 API
  version: 1.0.0
  description: |
    库存调拨规划器提供完整的调拨申请、库存约束检查、调拨方案生成与确认、多仓自动分配、
    仓库管理、SKU 管理、库存管理及调拨记录查询等能力。

    **主要模块：**
    - 仓库管理 (`/api/warehouses`)
    - SKU 管理 (`/api/skus`)
    - 库存管理 (`/api/inventory`)
    - 调拨申请 (`/api/transfer-requests`)
    - 调拨方案 (`/api/transfer-plans`)
    - 调拨记录 (`/api/transfer-records`)

    **约束分级：**
    - `FATAL`：不可通过，阻断调拨
    - `WARNING`：可确认但需业务复核
    - `INFO`：提示性信息

    **调拨方案状态机：**
    - `PENDING`（待确认）→ `CONFIRMED`（已确认）→ `COMPLETED`（已完成）
    - `PENDING` → `REJECTED`（已驳回）
    - 任意状态 → `CANCELLED`（已取消）
  contact:
    name: Inventory Ops Team
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: /
    description: 当前部署
...
paths:
...
  /api/transfer-plans:
    get:
      tags: [Transfer Plans]
      summary: 查询调拨方案列表
      operationId: listTransferPlans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REJECTED, PROCESSING, COMPLETED, CANCELLED]
        - name: request_id
          in: query
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize50'
      responses:
        '200': { description: 分页方案列表 }
...
  /api/transfer-plans/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Transfer Plans]
      summary: 查询方案详情（含明细 + 约束违规列表）
      operationId: getTransferPlan
      responses:
        '200': { description: 详情 }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Transfer Plans]
      summary: 删除调拨方案
      operationId: deleteTransferPlan
      description: 仅 PENDING/REJECTED/CANCELLED 可删除
      responses:
        '200': { description: 删除成功 }
  /api/transfer-plans/generate/{requestId}:
    post:
      tags: [Transfer Plans]
      summary: 生成调拨方案
      operationId: generateTransferPlan
      description: 根据申请自动生成方案；未指定源仓时触发多仓分配（中心仓优先+安全库存保护）
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                planner: { type: string, maxLength: 50, description: 规划人 }
      responses:
        '201': { description: 方案已生成 }
        '400':
          description: 申请状态不允许/明细为空
  /api/transfer-plans/{id}/confirm:
    post:
      tags: [Transfer Plans]
      summary: 确认调拨方案
      operationId: confirmTransferPlan
      description: 事务性处理：源仓扣减可用量加在途、目标仓加在途量、写入 transfer_records、状态变更 CONFIRMED
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                review_comment: { type: string, maxLength: 500 }
                operator: { type: string, maxLength: 50 }
      responses:
        '200': { description: 确认成功 }
  /api/transfer-plans/{id}/reject:
    post:
      tags: [Transfer Plans]
      summary: 驳回调拨方案
      operationId: rejectTransferPlan
      description: 仅 PENDING 状态可驳回；若申请无其他 PENDING 方案，申请变为 REJECTED
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [review_comment]
              properties:
                review_comment: { type: string, minLength: 1, maxLength: 500 }
                planner: { type: string, maxLength: 50 }
      responses:
        '200': { description: 驳回成功 }
  /api/transfer-plans/{id}/execute:
    post:
      tags: [Transfer Plans]
      summary: 执行入库（调拨完成）
      operationId: executeTransferPlan
      description: 源仓扣减在途、目标仓在途转可用；方案 → COMPLETED，申请 → COMPLETED
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                operator: { type: string, maxLength: 50 }
      responses:
        '200': { description: 入库完成 }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1, example: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
    PageSize50:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: 面向用户的错误消息 }
        code:
          type: string
          description: 稳定的机器可识别错误码
          example: WAREHOUSE_NOT_FOUND
        requestId:
          type: string
          description: 请求追踪 ID，用于日志关联
        timestamp: { type: string, format: date-time }
...